api/requirements.txt
flask
flask-cors
tensorflow==2.16.1
numpy
opencv-python
Pillow
mediapipe==0.10.14
waitress
